+++ plugin-utilities-plugin/src/test/kotlin/org/jetbrains/research/pluginUtilities/openRepository/OpenRepositoryTest.kt
@@ -1,23 +1,25 @@
-package org.jetbrains.research.pluginUtilities
+package org.jetbrains.research.pluginUtilities.openRepository
 
+import org.jetbrains.research.pluginUtilities.BuildSystem
+import org.jetbrains.research.pluginUtilities.preprocessing.AndroidSdkPreprocessing
+import org.jetbrains.research.pluginUtilities.preprocessing.Preprocessor
 import org.jetbrains.research.pluginUtilities.util.Extension
 import org.jetbrains.research.pluginUtilities.util.ParametrizedBaseTest
-import org.jetbrains.research.pluginUtilities.util.assertDirectoriesEqual
-import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import java.io.File
-import java.nio.file.Files
 
 private const val ANDROID_SDK_PATH = "/Users/Egor.Porsev/android_sdk/sdk"
 
 @RunWith(Parameterized::class)
 class OpenRepositoryTest :
-    ParametrizedBaseTest(getResourcesRootPath(::OpenRepositoryTest, "java_mock_projects")) {
+    ParametrizedBaseTest(getResourcesRootPath(::OpenRepositoryTest, "../java_mock_projects")) {
+
+    private val buildSystems = listOf(BuildSystem.Maven, BuildSystem.Gradle)
 
     private val preprocessor = Preprocessor(listOf(AndroidSdkPreprocessing(ANDROID_SDK_PATH)))
-    private val repositoryOpener = RepositoryOpener(preprocessor, javaBuildSystems)
+    private val repositoryOpener = RepositoryOpener(preprocessor, buildSystems)
 
     @JvmField
     @Parameterized.Parameter(0)
@@ -28,7 +30,7 @@ class OpenRepositoryTest :
         @Parameterized.Parameters(name = "{index}: {0}")
         fun getTestData() = getInAndOutArray(
             ::OpenRepositoryTest,
-            resourcesRootName = "java_mock_projects",
+            resourcesRootName = "../java_mock_projects",
             inExtension = Extension.EMPTY,
             outExtension = null
         )
